// src/soumu/blocks/datatable/datatable.type.tsx
export interface CellComponentProps<DataItem = any, CustomData = any> {
  dataItem: DataItem;
  dataKey?: string;
  customData?: CustomData;
  /**
   * Current data item index;
   */
  index: number;
}
// ...
export interface ColumnConfig<DataItem = any> {
  /**
   * Column name/label.
   */
  label: string;

  /**
   * Value must be a property key of `DataItem`.
   */
  dataKey?: keyof DataItem & string;

  /**
   * Whether this column must be displayed.
   *  _(User can use "setting display column" feature to change this value)._
   *
   * Default: true
   */
  isVisible?: boolean;

  /**
   * Define whether this column can be setting as hidden.
   * IMPORTANT: There is new logic, fixed column will not be able to setting as hidden.
   *
   * Default: false(user can setting this column as hidden column).
   * @see TableConfig.fixedColumnNumber
   */
  disabled?: boolean;

  /**
   * Specify component that will be used to display cell content.
   * By defaut, it display plain text, so there is no need to use this property
   * if you don't want to custom cell content.
   */
  component?: ComponentType<CellComponentProps<DataItem>>;

  /**
   * Custom data will be passed to component(`ColumnConfig.component`) as a property named `customData`.
   */
  customData?: any;

  isSorted?: any;

  align?: 'inherit' | 'left' | 'center' | 'right' | 'justify';

  width: string;
}